Various changes
User profile image finished. Adding the rating to the movies/tv shows. Fixing search bar bug and some styles changes

+++ paw-webapp/persistance/src/main/java/ar/edu/itba/paw/persistance/ReviewJdbcDao.java
@@ -20,7 +20,7 @@ public class ReviewJdbcDao implements ReviewDao{
                     resultSet.getLong("contentid"),
                     resultSet.getString("name"),
                     resultSet.getString("description"),
-                    resultSet.getLong("rating"),
+                    resultSet.getInt("rating"),
                     resultSet.getString("username"),
                     resultSet.getString("contentname"));
